Revert "slit"
This reverts commit 7b17243ce390b64df669ac001c0afb5a663a6696.
