+++ components/raftstore-v2/src/raft/peer.rs
@@ -65,7 +65,7 @@ pub struct Peer<EK: KvEngine, ER: RaftEngine> {
 
     /// Encoder for batching proposals and encoding them in a more efficient way
     /// than protobuf.
-    raw_write_encoder: Option<SimpleWriteReqEncoder>,
+    raw_write_encoder: Option<SimpleWriteReqEncoder<CmdResChannel>>,
     proposals: ProposalQueue<Vec<CmdResChannel>>,
     apply_scheduler: Option<ApplyScheduler>,
 
@@ -614,12 +614,14 @@ impl<EK: KvEngine, ER: RaftEngine> Peer<EK, ER> {
     }
 
     #[inline]
-    pub fn simple_write_encoder_mut(&mut self) -> &mut Option<SimpleWriteReqEncoder> {
+    pub fn simple_write_encoder_mut(
+        &mut self,
+    ) -> &mut Option<SimpleWriteReqEncoder<CmdResChannel>> {
         &mut self.raw_write_encoder
     }
 
     #[inline]
-    pub fn simple_write_encoder(&self) -> &Option<SimpleWriteReqEncoder> {
+    pub fn simple_write_encoder(&self) -> &Option<SimpleWriteReqEncoder<CmdResChannel>> {
         &self.raw_write_encoder
     }
