+++ components/raftstore/src/store/message/simple_write.rs
@@ -12,6 +12,7 @@ use slog::Logger;
 use tikv_util::slog_panic;
 
 use super::parse_at;
+use crate::store::{msg::ErrorCallback, WriteCallback};
 
 // MAGIC number to hint simple write codec is used. If it's a protobuf message,
 // the first one or several bytes are for field tag, which can't be zero.
@@ -40,14 +41,22 @@ impl SimpleWriteBinary {
 /// not efficient enough for simple request. `SimpleWrite` is introduce to
 /// make codec alloc less and fast.
 #[derive(Debug)]
-pub struct SimpleWriteReqEncoder {
+pub struct SimpleWriteReqEncoder<C>
+where
+    C: ErrorCallback + WriteCallback,
+{
     header: Box<RaftRequestHeader>,
     buf: Vec<u8>,
+    channels: Vec<C>,
     size_limit: usize,
     write_type: WriteType,
+    notify_proposed: bool,
 }
 
-impl SimpleWriteReqEncoder {
+impl<C> SimpleWriteReqEncoder<C>
+where
+    C: ErrorCallback + WriteCallback,
+{
     /// Create a request encoder.
     ///
     /// If `notify_proposed` is true, channels will be called `notify_proposed`
@@ -56,7 +65,8 @@ impl SimpleWriteReqEncoder {
         header: Box<RaftRequestHeader>,
         bin: SimpleWriteBinary,
         size_limit: usize,
-    ) -> SimpleWriteReqEncoder {
+        notify_proposed: bool,
+    ) -> SimpleWriteReqEncoder<C> {
         let mut buf = Vec::with_capacity(256);
         buf.push(MAGIC_PREFIX);
         header.write_length_delimited_to_vec(&mut buf).unwrap();
@@ -65,8 +75,10 @@ impl SimpleWriteReqEncoder {
         SimpleWriteReqEncoder {
             header,
             buf,
+            channels: vec![],
             size_limit,
             write_type: bin.write_type,
+            notify_proposed,
         }
     }
 
@@ -96,8 +108,21 @@ impl SimpleWriteReqEncoder {
     }
 
     #[inline]
-    pub fn encode(self) -> Vec<u8> {
-        self.buf
+    pub fn encode(self) -> (Vec<u8>, Vec<C>) {
+        (self.buf, self.channels)
+    }
+
+    #[inline]
+    pub fn add_response_channel(&mut self, mut ch: C) {
+        if self.notify_proposed {
+            ch.notify_proposed();
+        }
+        self.channels.push(ch);
+    }
+
+    #[inline]
+    pub fn notify_proposed(&self) -> bool {
+        self.notify_proposed
     }
 
     #[inline]
